Fix EAP build
