+++ components/ide/jetbrains/gateway-plugin/gradle-latest.properties
@@ -8,4 +8,4 @@ pluginUntilBuild=243.*
 # See https://jb.gg/intellij-platform-builds-list for available build versions.
 pluginVerifierIdeVersions=2024.3
 # Version from "com.jetbrains.gateway" which can be found at https://www.jetbrains.com/intellij-repository/snapshots
-platformVersion=2024.3
+platformVersion=243.18137.22
