+++ components/ide/jetbrains/gateway-plugin/src/main/kotlin/io/gitpod/jetbrains/gateway/common/GitpodConnectionHandleFactory.kt
@@ -1,0 +1,41 @@
+package io.gitpod.jetbrains.gateway.common
+
+import com.jetbrains.gateway.api.CustomConnectionFrameComponentProvider
+import com.jetbrains.gateway.api.CustomConnectionFrameContext
+import com.jetbrains.gateway.api.GatewayConnectionHandle
+import com.jetbrains.gateway.ssh.HostTunnelConnector
+import com.jetbrains.gateway.thinClientLink.ThinClientHandle
+import com.jetbrains.rd.util.lifetime.Lifetime
+import io.gitpod.jetbrains.gateway.GitpodConnectionProvider.ConnectParams
+import java.net.URI
+import javax.swing.JComponent
+
+@Suppress("UnstableApiUsage")
+interface GitpodConnectionHandleFactory {
+    fun createGitpodConnectionHandle(
+        lifetime: Lifetime,
+        component: JComponent,
+        params: ConnectParams
+    ): GatewayConnectionHandle
+
+    suspend fun connect(lifetime: Lifetime, connector: HostTunnelConnector, tcpJoinLink: URI): ThinClientHandle
+}
+
+class GitpodConnectionHandle(
+    lifetime: Lifetime,
+    private val component: JComponent,
+    private val params: ConnectParams
+) : GatewayConnectionHandle(lifetime) {
+    override fun customComponentProvider(lifetime: Lifetime) = object : CustomConnectionFrameComponentProvider {
+        override val closeConfirmationText = "Disconnect from ${getTitle()}?"
+        override fun createComponent(context: CustomConnectionFrameContext) = component
+    }
+
+    override fun getTitle(): String {
+        return params.title
+    }
+
+    override fun hideToTrayOnStart(): Boolean {
+        return false
+    }
+}
