+++ components/ide/jetbrains/gateway-plugin/src/main/resources-latest/META-INF/extensions.xml
@@ -5,6 +5,8 @@
 -->
 <idea-plugin>
     <extensions defaultExtensionNs="com.intellij">
+        <applicationService serviceInterface="io.gitpod.jetbrains.gateway.common.GitpodConnectionHandleFactory"
+                            serviceImplementation="io.gitpod.jetbrains.gateway.latest.LatestGitpodConnectionHandleFactory"/>
     </extensions>
     <extensions defaultExtensionNs="com.jetbrains">
     </extensions>
